200 seconds so close!

+++ tessellate/tessellate.py
@@ -1429,10 +1429,12 @@ python {self.working_path}/cubing_scripts/S{self.sector}C{cam}C{ccd}_script.py"
                         python_text = f"\
 from tessellate import DataProcessor\n\
 \n\
+split = {self.split}\n\
 processor = DataProcessor(sector={self.sector},path='{self.data_path}',verbose=2)\n\
-processor.make_cuts(cam={cam},ccd={ccd},n={self.n},cut={cut},split={self.split})\n\
-with open(f'{self.data_path}/Sector{self.sector}/Cam{cam}/Ccd{ccd}/Cut{cut}of{self.n**2}/cut.txt', 'w') as file:\n\
-    file.write('Cut!')"
+processor.make_cuts(cam={cam},ccd={ccd},n={self.n},cut={cut},split=split)\n\
+if not split:\n\
+    with open(f'{self.data_path}/Sector{self.sector}/Cam{cam}/Ccd{ccd}/Cut{cut}of{self.n**2}/cut.txt', 'w') as file:\n\
+        file.write('Cut!')"
 
                         with open(f"{self.working_path}/cutting_scripts/S{self.sector}C{cam}C{ccd}C{cut}_script.py", "w") as python_file:
                             python_file.write(python_text)
@@ -1513,11 +1515,13 @@ python {self.working_path}/cutting_scripts/S{self.sector}C{cam}C{ccd}C{cut}_scri
 from tessellate import DataProcessor\n\
 import os\n\
 \n\
+split={self.split}\n\
 processor = DataProcessor(sector={self.sector},path='{self.data_path}',verbose=2)\n\
-processor.reduce(cam={cam},ccd={ccd},n={self.n},cut={cut})\n\
-if os.path.exists('{self.data_path}/Sector{self.sector}/Cam{cam}/Ccd{ccd}/Cut{cut}of{self.n**2}/sector{self.sector}_cam{cam}_ccd{ccd}_cut{cut}_of{self.n**2}_Shifts.npy'):\n\
-    with open(f'{self.data_path}/Sector{self.sector}/Cam{cam}/Ccd{ccd}/Cut{cut}of{self.n**2}/reduced.txt', 'w') as file:\n\
-        file.write('Reduced with TESSreduce version {tr.__version__}.')"
+processor.reduce(cam={cam},ccd={ccd},n={self.n},cut={cut},split=split)\n\
+if not split:\n\
+    if os.path.exists('{self.data_path}/Sector{self.sector}/Cam{cam}/Ccd{ccd}/Cut{cut}of{self.n**2}/sector{self.sector}_cam{cam}_ccd{ccd}_cut{cut}_of{self.n**2}_Shifts.npy'):\n\
+        with open(f'{self.data_path}/Sector{self.sector}/Cam{cam}/Ccd{ccd}/Cut{cut}of{self.n**2}/reduced.txt', 'w') as file:\n\
+            file.write('Reduced with TESSreduce version {tr.__version__}.')"
                 
                         with open(f"{self.working_path}/reduction_scripts/S{self.sector}C{cam}C{ccd}C{cut}_script.py", "w") as python_file:
                             python_file.write(python_text)
